Add operations based on selected text via "ll" locator
You can type "Ctrl+K ll " to get a list of predefined prompts like:

 * Create a summary
 * Create a commit message
 * Explain this code
 * Generate test cases. Output only code. No explanations

But you can type your own prompt, which will be saved as part of a 50
items "History".

When the locator entry is executed, the a new llama.cpp Conversation is
open with the selected text passed with the selected prompt.