Revisited the message generation / regeneration
The message chat is now properly filtered by the leafNodeId, just like in
the original React WebUI application.

This means that now you can edit a message, and you will get branches of
the conversations.

Similarly for the regeneration of the answers.

Move the ChatEditor class declaration to the header, for better
consistency with the rest.

Copied the comment that describes chat branching to llamatypes.h.

Added a ollama like spinner as a label for the thinking part, when the
thought process is not expanded.